Add build option to compile only

+++ scripts/build-vagrant.sh
@@ -20,19 +20,21 @@ mkdir $DEBUG_BUILD &&
 sleep 1
 
 dostuff 'cd /home/libnss-maria/Debug && cmake -D CMAKE_BUILD_TYPE=Debug .. \
-&& make && ctest --verbose' &&
+&& make && ctest --verbose'
 
-# go to scripts directory, ugly, but well..
-cd $SCRIPTS_DIR &&
+if [[ $? -eq 0 && -z $COMPILE_ONLY ]]; then
+  # go to scripts directory, ugly, but well..
+  cd $SCRIPTS_DIR &&
 
-dostuff 'sudo rm /lib/libnss_maria.so*'
-dostuff 'sudo cp /home/libnss-maria/Debug/src/libnss_maria.so.2.0.0 /lib/libnss_maria.so.2.0.0' &&
-dostuff 'sudo ln -s /lib/libnss_maria.so.2.0.0 /lib/libnss_maria.so' &&
-dostuff 'sudo ln -s /lib/libnss_maria.so.2.0.0 /lib/libnss_maria.so.2' &&
-dostuff 'sudo chmod 644 /lib/libnss_maria.so.2.0.0' &&
+  dostuff 'sudo rm /lib/libnss_maria.so*'
+  dostuff 'sudo cp /home/libnss-maria/Debug/src/libnss_maria.so.2.0.0 /lib/libnss_maria.so.2.0.0' &&
+  dostuff 'sudo ln -s /lib/libnss_maria.so.2.0.0 /lib/libnss_maria.so' &&
+  dostuff 'sudo ln -s /lib/libnss_maria.so.2.0.0 /lib/libnss_maria.so.2' &&
+  dostuff 'sudo chmod 644 /lib/libnss_maria.so.2.0.0' &&
 
-dostuff 'sudo cp /home/libnss-maria/examples/sos-sso/nsswitch.conf /etc' &&
-dostuff 'sudo cp /home/libnss-maria/examples/sos-sso/libnss-maria.conf /etc' &&
-dostuff 'sudo cp /home/libnss-maria/examples/sos-sso/libnss-maria-root.conf /etc' &&
+  dostuff 'sudo cp /home/libnss-maria/examples/sos-sso/nsswitch.conf /etc' &&
+  dostuff 'sudo cp /home/libnss-maria/examples/sos-sso/libnss-maria.conf /etc' &&
+  dostuff 'sudo cp /home/libnss-maria/examples/sos-sso/libnss-maria-root.conf /etc' &&
 
-dostuff 'getent passwd katarina || echo "No user found"'
+  dostuff 'getent passwd katarina || echo "No user found"'
+fi
